chore: move blog post schemas to blog folder
