+++ apps/cms/schemas/blog/defaultSeo.ts
