+++ apps/cms/schemas/blog/rss.ts
